refactor: flatten component structure with co-located CSS
- Move components to flat structure (nav-bar.tsx + nav-bar.css)
- Remove nested navigation/ and content/ directories
- Add co-located CSS files for each component following VitePress SFC pattern
- Update imports to use simplified flat component paths
- Move layout-specific styles to doc-layout.css
- Simplify component organization for better maintainability
