+++ src/lib/components/nav-bar.css
@@ -1,0 +1,35 @@
+/* Navigation Bar */
+.VPNavBar {
+  position: relative;
+  height: var(--vp-nav-height);
+  background-color: var(--vp-nav-bg-color);
+  border-bottom: 1px solid var(--vp-c-divider);
+  transition: background-color 0.25s;
+}
+
+@media (min-width: 960px) {
+  .VPNavBar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    z-index: var(--vp-z-index-nav);
+  }
+}
+
+.VPNavBar .container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 auto;
+  max-width: calc(var(--vp-layout-max-width) - 64px);
+  height: var(--vp-nav-height);
+  padding: 0 32px 0 24px;
+}
+
+.VPNavBar .title {
+  font-size: 16px;
+  font-weight: 600;
+  color: var(--vp-c-text-1);
+  flex-shrink: 0;
+}
