+++ src/lib/components/sidebar.css
@@ -1,0 +1,68 @@
+/* Sidebar */
+.VPSidebar {
+  position: fixed;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  z-index: var(--vp-z-index-sidebar);
+  padding: 32px 32px 96px;
+  width: calc(100vw - 64px);
+  max-width: 320px;
+  background-color: var(--vp-sidebar-bg-color);
+  opacity: 0;
+  transform: translateX(-100%);
+  transition: opacity 0.5s, transform 0.25s ease;
+}
+
+@media (min-width: 960px) {
+  .VPSidebar {
+    padding-top: calc(var(--vp-nav-height) + 32px);
+    width: var(--vp-sidebar-width);
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+
+.VPSidebar.open {
+  opacity: 1;
+  transform: translateX(0);
+}
+
+.VPSidebar nav {
+  outline: none;
+}
+
+.sidebar-group h5 {
+  margin: 0 0 8px 0;
+  font-size: 13px;
+  font-weight: 600;
+  color: var(--vp-c-text-2);
+  text-transform: uppercase;
+  letter-spacing: 0.08em;
+}
+
+.sidebar-links {
+  list-style: none;
+  padding: 0;
+  margin: 0 0 24px 0;
+}
+
+.sidebar-link {
+  display: block;
+  line-height: 32px;
+  font-size: 14px;
+  font-weight: 400;
+  color: var(--vp-c-text-2);
+  white-space: nowrap;
+  text-decoration: none;
+  transition: color 0.25s;
+}
+
+.sidebar-link:hover {
+  color: var(--vp-c-text-1);
+}
+
+.sidebar-link.active {
+  color: var(--vp-c-brand-1);
+  font-weight: 600;
+}
