+++ src/lib/components/sidebar.tsx
@@ -1,3 +1,5 @@
+import './sidebar.css';
+
 export function Sidebar() {
 	return (
 		<aside className="VPSidebar">
