+++ src/lib/layout/doc-layout.css
@@ -1,0 +1,127 @@
+/* Layout Variables */
+:root {
+  --vp-layout-max-width: 1440px;
+  --vp-nav-height: 64px;
+  --vp-sidebar-width: 272px;
+  --vp-z-index-nav: 30;
+  --vp-z-index-sidebar: 60;
+  --vp-nav-bg-color: var(--vp-c-bg);
+  --vp-sidebar-bg-color: var(--vp-c-bg-alt);
+}
+
+@media (min-width: 960px) {
+  :root {
+    --vp-z-index-sidebar: 25;
+  }
+}
+
+/* Main Layout */
+.Layout {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+/* Content Area */
+.VPContent {
+  flex-grow: 1;
+  flex-shrink: 0;
+  margin: 0 auto;
+  width: 100%;
+}
+
+.VPContent.has-sidebar {
+  margin: 0;
+}
+
+@media (min-width: 960px) {
+  .VPContent {
+    padding-top: var(--vp-nav-height);
+  }
+  
+  .VPContent.has-sidebar {
+    padding-left: var(--vp-sidebar-width);
+  }
+}
+
+@media (min-width: 1440px) {
+  .VPContent.has-sidebar {
+    padding-right: calc((100vw - var(--vp-layout-max-width)) / 2);
+    padding-left: calc((100vw - var(--vp-layout-max-width)) / 2 + var(--vp-sidebar-width));
+  }
+}
+
+/* Doc Layout */
+.VPDoc {
+  padding: 32px 24px 96px;
+  width: 100%;
+}
+
+@media (min-width: 768px) {
+  .VPDoc {
+    padding: 48px 32px 128px;
+  }
+}
+
+@media (min-width: 1280px) {
+  .VPDoc .container {
+    display: flex;
+    justify-content: center;
+  }
+  
+  .VPDoc .aside {
+    display: block;
+  }
+}
+
+.VPDoc .container {
+  margin: 0 auto;
+  width: 100%;
+}
+
+.VPDoc .content {
+  position: relative;
+  margin: 0 auto;
+  width: 100%;
+}
+
+@media (min-width: 1280px) {
+  .VPDoc .content {
+    order: 1;
+    margin: 0;
+    min-width: 640px;
+  }
+}
+
+.VPDoc.has-aside .content-container {
+  max-width: 688px;
+}
+
+.VPDoc .content-container {
+  max-width: 688px;
+}
+
+/* Responsive backdrop */
+.backdrop {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  z-index: var(--vp-z-index-sidebar);
+  background: rgba(0, 0, 0, 0.6);
+  opacity: 0;
+  visibility: hidden;
+  transition: opacity 0.25s, visibility 0.25s;
+}
+
+.backdrop.show {
+  opacity: 1;
+  visibility: visible;
+}
+
+@media (min-width: 960px) {
+  .backdrop {
+    display: none;
+  }
+}
