+++ src/lib/styles/index.css
@@ -1,4 +1,3 @@
 @import './tokens.css';
 @import './base.css';
-@import './layout.css';
 @import './vp-doc.css';
