--- src/lib/styles/layout.css
@@ -1,306 +1,0 @@
-/* VitePress Layout Variables */
-:root {
-  --vp-layout-max-width: 1440px;
-  --vp-nav-height: 64px;
-  --vp-sidebar-width: 272px;
-  --vp-z-index-nav: 30;
-  --vp-z-index-sidebar: 60;
-  --vp-nav-bg-color: var(--vp-c-bg);
-  --vp-sidebar-bg-color: var(--vp-c-bg-alt);
-}
-
-@media (min-width: 960px) {
-  :root {
-    --vp-z-index-sidebar: 25;
-  }
-}
-
-/* Main Layout */
-.Layout {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
-
-/* Navigation Bar */
-.VPNavBar {
-  position: relative;
-  height: var(--vp-nav-height);
-  background-color: var(--vp-nav-bg-color);
-  border-bottom: 1px solid var(--vp-c-divider);
-  transition: background-color 0.25s;
-}
-
-@media (min-width: 960px) {
-  .VPNavBar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    z-index: var(--vp-z-index-nav);
-  }
-}
-
-.VPNavBar .container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 0 auto;
-  max-width: calc(var(--vp-layout-max-width) - 64px);
-  height: var(--vp-nav-height);
-  padding: 0 32px 0 24px;
-}
-
-.VPNavBar .title {
-  font-size: 16px;
-  font-weight: 600;
-  color: var(--vp-c-text-1);
-  flex-shrink: 0;
-}
-
-/* Sidebar */
-.VPSidebar {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  z-index: var(--vp-z-index-sidebar);
-  padding: 32px 32px 96px;
-  width: calc(100vw - 64px);
-  max-width: 320px;
-  background-color: var(--vp-sidebar-bg-color);
-  opacity: 0;
-  transform: translateX(-100%);
-  transition: opacity 0.5s, transform 0.25s ease;
-}
-
-@media (min-width: 960px) {
-  .VPSidebar {
-    padding-top: calc(var(--vp-nav-height) + 32px);
-    width: var(--vp-sidebar-width);
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-.VPSidebar.open {
-  opacity: 1;
-  transform: translateX(0);
-}
-
-.VPSidebar nav {
-  outline: none;
-}
-
-.sidebar-group h5 {
-  margin: 0 0 8px 0;
-  font-size: 13px;
-  font-weight: 600;
-  color: var(--vp-c-text-2);
-  text-transform: uppercase;
-  letter-spacing: 0.08em;
-}
-
-.sidebar-links {
-  list-style: none;
-  padding: 0;
-  margin: 0 0 24px 0;
-}
-
-.sidebar-link {
-  display: block;
-  line-height: 32px;
-  font-size: 14px;
-  font-weight: 400;
-  color: var(--vp-c-text-2);
-  white-space: nowrap;
-  text-decoration: none;
-  transition: color 0.25s;
-}
-
-.sidebar-link:hover {
-  color: var(--vp-c-text-1);
-}
-
-.sidebar-link.active {
-  color: var(--vp-c-brand-1);
-  font-weight: 600;
-}
-
-/* Content Area */
-.VPContent {
-  flex-grow: 1;
-  flex-shrink: 0;
-  margin: 0 auto;
-  width: 100%;
-}
-
-.VPContent.has-sidebar {
-  margin: 0;
-}
-
-@media (min-width: 960px) {
-  .VPContent {
-    padding-top: var(--vp-nav-height);
-  }
-  
-  .VPContent.has-sidebar {
-    padding-left: var(--vp-sidebar-width);
-  }
-}
-
-@media (min-width: 1440px) {
-  .VPContent.has-sidebar {
-    padding-right: calc((100vw - var(--vp-layout-max-width)) / 2);
-    padding-left: calc((100vw - var(--vp-layout-max-width)) / 2 + var(--vp-sidebar-width));
-  }
-}
-
-/* Doc Layout */
-.VPDoc {
-  padding: 32px 24px 96px;
-  width: 100%;
-}
-
-@media (min-width: 768px) {
-  .VPDoc {
-    padding: 48px 32px 128px;
-  }
-}
-
-@media (min-width: 1280px) {
-  .VPDoc .container {
-    display: flex;
-    justify-content: center;
-  }
-  
-  .VPDoc .aside {
-    display: block;
-  }
-}
-
-.VPDoc .container {
-  margin: 0 auto;
-  width: 100%;
-}
-
-.VPDoc .content {
-  position: relative;
-  margin: 0 auto;
-  width: 100%;
-}
-
-@media (min-width: 1280px) {
-  .VPDoc .content {
-    order: 1;
-    margin: 0;
-    min-width: 640px;
-  }
-}
-
-.VPDoc.has-aside .content-container {
-  max-width: 688px;
-}
-
-.VPDoc .content-container {
-  max-width: 688px;
-}
-
-/* Table of Contents (Aside) */
-.aside {
-  position: relative;
-  display: none;
-  order: 2;
-  flex-grow: 1;
-  padding-left: 32px;
-  width: 100%;
-  max-width: 256px;
-}
-
-@media (min-width: 1280px) {
-  .aside {
-    display: block;
-  }
-}
-
-.aside-container {
-  position: fixed;
-  top: 0;
-  padding-top: calc(var(--vp-nav-height) + 48px);
-  width: 224px;
-  height: 100vh;
-  overflow-x: hidden;
-  overflow-y: auto;
-}
-
-.aside-content {
-  display: flex;
-  flex-direction: column;
-  min-height: calc(100vh - (var(--vp-nav-height) + 48px));
-  padding-bottom: 32px;
-}
-
-.outline-title {
-  font-size: 13px;
-  font-weight: 600;
-  color: var(--vp-c-text-2);
-  text-transform: uppercase;
-  letter-spacing: 0.08em;
-  margin-bottom: 12px;
-}
-
-.outline-list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.outline-item {
-  line-height: 28px;
-}
-
-.outline-link {
-  display: block;
-  font-size: 13px;
-  font-weight: 400;
-  color: var(--vp-c-text-2);
-  text-decoration: none;
-  transition: color 0.25s;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.outline-link:hover {
-  color: var(--vp-c-text-1);
-}
-
-.outline-link.active {
-  color: var(--vp-c-brand-1);
-  font-weight: 600;
-}
-
-/* Responsive backdrop */
-.backdrop {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: var(--vp-z-index-sidebar);
-  background: rgba(0, 0, 0, 0.6);
-  opacity: 0;
-  visibility: hidden;
-  transition: opacity 0.25s, visibility 0.25s;
-}
-
-.backdrop.show {
-  opacity: 1;
-  visibility: visible;
-}
-
-@media (min-width: 960px) {
-  .backdrop {
-    display: none;
-  }
-}
